+++ djpcms/utils/html/__init__.py
@@ -1,6 +1,5 @@
 from pagination import *
 from base import *
-from form import *
 from grid960 import *
 from box import *
 from autocomplete import *